requirements.txt
beautifulsoup4==4.11.2
Django==4.1.1
djangorestframework==3.14.0
selenium==4.8.0
python-amazon-paapi==5.0.1
requests==2.28.2
whitenoise==6.3.0
